refactor: 코드리뷰 반영
- memberService -> coffeeChatService로 이동

related issue #548
