+++ src/main/java/org/sopt/makers/internal/controller/InternalOpenApiController.java
@@ -23,6 +23,7 @@ import org.sopt.makers.internal.member.controller.coffeechat.dto.response.Intern
 import org.sopt.makers.internal.dto.internal.InternalMemberProjectResponse;
 import org.sopt.makers.internal.member.mapper.coffeechat.CoffeeChatResponseMapper;
 import org.sopt.makers.internal.member.repository.coffeechat.dto.InternalCoffeeChatMemberDto;
+import org.sopt.makers.internal.member.service.coffeechat.CoffeeChatService;
 import org.sopt.makers.internal.service.InternalApiService;
 import org.sopt.makers.internal.service.MemberService;
 import org.sopt.makers.internal.service.ProjectService;
@@ -50,6 +51,8 @@ public class InternalOpenApiController {
     private final InternalApiService internalApiService;
     private final MemberService memberService;
     private final MemberMapper memberMapper;
+
+    private final CoffeeChatService coffeeChatService;
     private final CoffeeChatResponseMapper coffeeChatResponseMapper;
 
     private final ProjectService projectService;
@@ -328,7 +331,7 @@ public class InternalOpenApiController {
     @Operation(summary = "커피챗 오픈 유저 리스트 조회 API")
     @GetMapping("/members/coffeechat")
     public ResponseEntity<List<InternalCoffeeChatMemberResponse>> getCoffeeChatActivateMembers() {
-        List<InternalCoffeeChatMemberDto> members = memberService.getCoffeeChatActivateMembers();
+        List<InternalCoffeeChatMemberDto> members = coffeeChatService.getCoffeeChatActivateMembers();
         List<InternalCoffeeChatMemberResponse> response = coffeeChatResponseMapper.toInternalCoffeeChatMemberResponse(members);
         return ResponseEntity.status(HttpStatus.OK).body(response);
     }
