+++ src/main/java/org/sopt/makers/internal/service/MemberService.java
@@ -210,14 +210,6 @@ public class MemberService {
         else return team;
     }
 
-    @Transactional(readOnly = true)
-    public List<InternalCoffeeChatMemberDto> getCoffeeChatActivateMembers() {
-        List<Member> members = memberProfileQueryRepository.findAllMembersByCoffeeChatActivate();
-        return members.stream().map(member -> InternalCoffeeChatMemberDto.of(
-                member, memberRetriever.concatPartAndGeneration(member.getId())
-        )).toList();
-    }
-
     @Transactional
     public Member saveMemberProfile(Long id, MemberProfileSaveRequest request) {
         val member = memberRepository.findById(id).orElseThrow(() -> new NotFoundDBEntityException("Member"));
